requirements.txt
django==5.2.7
djangorestframework==3.16.1
mysqlclient==2.2.7
sentence-transformers==5.1.2
# added our custom library
mariax==1.1.0